+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/api/LoggingBase.scala
@@ -31,7 +31,6 @@ package com.tersesystems.echopraxia.plusscala.api
  */
 trait LoggingBase
     extends ValueTypeClasses
-    with ToValueAttributeImplicits
     with FieldConversionImplicits
     with OptionValueTypes
     with EitherValueTypes
